requirements.txt
discord.py[voice]==2.7.1
yt-dlp==2026.3.17
python-dotenv==1.2.2
psycopg2-binary==2.9.9